+++ libsrc/leddevice/dev_spi/LedDeviceAPA102.cpp
@@ -33,7 +33,7 @@ bool LedDeviceAPA102::init(const QJsonObject &deviceConfig)
 
 int LedDeviceAPA102::write(const std::vector<ColorRgb> &ledValues)
 {
-	for (signed iLed=0; iLed < _ledCount; ++iLed) {
+	for (signed iLed=0; iLed < static_cast<int>( _ledCount); ++iLed) {
 		const ColorRgb& rgb = ledValues[iLed];
 		_ledBuffer[4+iLed*4]   = 0xFF;
 		_ledBuffer[4+iLed*4+1] = rgb.red;
